Performs Canonical Correlation Analysis using Empirical Orthogonal Functions from a lagged dataset using a stars object

Performs CCA using EOF analysis using a stars object

Usage

# S3 method for stars
cancor_eof(x, lag = 7, n_eof = 10, ...)

Arguments

x

The dataframe or stars object. If it is a dataframe, then it should have the locations.

lag

Specifies the lag to be used.

n_eof

The number of EOFs to be used.

...

Other arguments that need to be used for datafames or currently ignored.

Examples

library(dplyr)
library(stars)
# Create a stars object from a data frame
precip_df <- NOAA_df_1990[NOAA_df_1990$proc == 'Precip', ] %>%
   filter(date >= "1992-02-01" & date <= "1992-02-28")
precip <- precip_df[ ,c('lat', 'lon', 'date', 'z')]
st_precip <- st_as_stars(precip, dims = c("lon", "lat", "date"))
cancor_eof(st_precip)
#> [1] "Canonical Correlation Analysis is carried out using a lagged dataset."
#> $cancor_df
#>     t          ts Variable
#> 1   1  0.09854253     CCA1
#> 2   2  0.09980185     CCA1
#> 3   3  0.09699787     CCA1
#> 4   4  0.08426217     CCA1
#> 5   5 -0.11499212     CCA1
#> 6   6 -0.04770059     CCA1
#> 7   7  0.11190785     CCA1
#> 8   8  0.11082335     CCA1
#> 9   9  0.09869250     CCA1
#> 10 10  0.08272116     CCA1
#> 11 11  0.12315778     CCA1
#> 12 12  0.11295216     CCA1
#> 13 13  0.09776756     CCA1
#> 14 14  0.05611574     CCA1
#> 15 15 -0.05992945     CCA1
#> 16 16  0.10522065     CCA1
#> 17 17 -0.89912063     CCA1
#> 18 18 -0.13212227     CCA1
#> 19 19  0.11432431     CCA1
#> 20 20  0.09659169     CCA1
#> 21 21  0.09214651     CCA1
#> 22  1  0.09031854     CCA2
#> 23  2  0.08967483     CCA2
#> 24  3  0.08795219     CCA2
#> 25  4  0.07492675     CCA2
#> 26  5 -0.12432863     CCA2
#> 27  6 -0.05697822     CCA2
#> 28  7  0.10256711     CCA2
#> 29  8  0.10148593     CCA2
#> 30  9  0.08922849     CCA2
#> 31 10  0.07339788     CCA2
#> 32 11  0.11382668     CCA2
#> 33 12  0.10364830     CCA2
#> 34 13  0.08841009     CCA2
#> 35 14  0.04678376     CCA2
#> 36 15 -0.06925361     CCA2
#> 37 16  0.09588205     CCA2
#> 38 17 -0.90845080     CCA2
#> 39 18 -0.14145409     CCA2
#> 40 19  0.10508777     CCA2
#> 41 20  0.08696339     CCA2
#> 42 21  0.08258226     CCA2
#> 
#> $eofs_df
#>           lon      lat           SW1           SW2
#> 1   -95.37083 45.86666 -3.334517e-04  7.336513e-05
#> 2   -94.12500 45.65000 -6.407051e-04 -7.861744e-03
#> 3   -84.79583 45.54167 -7.280403e-03 -7.530529e-05
#> 4   -98.42500 45.38333 -9.158982e-03  2.238351e-03
#> 5   -83.52500 45.12500  2.138998e-04  5.878987e-03
#> 6   -89.62083 44.94583 -5.955375e-03 -1.978085e-02
#> 7   -93.20000 44.88750 -4.390209e-03 -1.513789e-02
#> 8   -91.48750 44.83750 -2.464208e-03 -1.782965e-02
#> 9   -85.60417 44.72083 -7.095144e-04  3.819908e-03
#> 10  -95.06667 44.57917 -6.116133e-03 -2.647446e-03
#> 11  -88.22917 44.47500 -2.234870e-03 -6.908015e-03
#> 12  -98.23333 44.40417 -1.809035e-02  4.104889e-03
#> 13  -84.70417 44.27917 -2.462678e-03  6.969687e-04
#> 14  -92.60417 44.07500 -4.726942e-03 -3.830219e-03
#> 15  -91.31250 43.85000 -5.713556e-03 -2.427281e-03
#> 16  -96.68750 43.54167 -2.541550e-02 -2.656180e-04
#> 17  -84.08750 43.54167  1.004889e-04 -4.770875e-03
#> 18  -86.14583 43.25417  1.091295e-03 -7.416754e-03
#> 19  -93.39167 43.15000 -1.221748e-02  1.124030e-03
#> 20  -89.33333 43.09583  2.015151e-03 -9.062545e-03
#> 21  -83.73333 43.00417  4.751884e-03  5.942778e-03
#> 22  -87.88750 42.93750 -7.052015e-03 -2.525619e-03
#> 23  -85.45834 42.87500  4.990834e-03 -4.051020e-03
#> 24  -84.58333 42.75000  5.825665e-03  4.109319e-03
#> 25  -92.37917 42.57083 -1.308627e-02 -4.073597e-03
#> 26  -90.81250 42.43333 -6.791653e-03  1.064387e-03
#> 27  -96.37917 42.35833 -7.374838e-03  1.843432e-03
#> 28  -84.49167 42.28750  5.824586e-03  3.868369e-03
#> 29  -83.26667 42.23333  8.649540e-03  6.226048e-03
#> 30  -89.11250 42.17917 -6.382062e-03 -1.050751e-02
#> 31  -80.15417 42.08750  8.596859e-03  4.991163e-03
#> 32  -97.38333 41.98333 -1.148741e-02  3.859388e-03
#> 33  -87.88750 41.98333  2.798308e-03 -1.111479e-02
#> 34  -91.73333 41.86250 -1.929923e-02  1.618593e-03
#> 35  -86.32500 41.72083  5.958925e-03  4.321448e-03
#> 36  -83.82917 41.60833  7.679068e-03  6.790043e-03
#> 37  -93.60000 41.52917 -1.136306e-02  9.940088e-03
#> 38  -90.51667 41.45833 -9.796473e-03 -1.389690e-02
#> 39  -81.80000 41.40417  3.220598e-03  1.196090e-02
#> 40  -96.07500 41.35833 -1.925637e-02  3.430410e-03
#> 41  -95.89584 41.30833 -2.156916e-02  2.312254e-03
#> 42  -80.62083 41.23333  5.840900e-03  2.975849e-03
#> 43  -92.60417 41.12083 -1.827082e-02  2.976565e-03
#> 44  -83.67916 41.03333  2.102201e-03  4.528252e-03
#> 45  -85.28750 40.99583  1.263316e-02  6.258655e-03
#> 46  -98.32083 40.96250 -2.129583e-02  5.361435e-03
#> 47  -81.46250 40.91250  6.060019e-03  1.273216e-03
#> 48  -96.68750 40.85833 -1.217387e-02  1.859042e-03
#> 49  -82.51667 40.78750  7.708463e-03  9.662005e-03
#> 50  -89.74167 40.66250 -1.934171e-03 -8.623908e-03
#> 51  -80.32083 40.41250  7.015192e-03 -1.053840e-02
#> 52  -82.90000 40.10417  3.378886e-03 -2.158662e-04
#> 53  -81.90417 39.95417  1.118433e-02  2.032598e-04
#> 54  -91.09583 39.92917 -2.713316e-02 -1.090280e-02
#> 55  -84.17916 39.87083 -1.598283e-03  4.525397e-03
#> 56  -89.74167 39.78333  8.789322e-03 -2.655208e-04
#> 57  -86.27083 39.69583 -7.317331e-05  5.548021e-03
#> 58  -97.77917 39.55417 -2.305458e-04  1.721297e-03
#> 59  -81.46250 39.39167  3.151998e-03 -4.312226e-03
#> 60  -94.77500 39.26667 -4.445478e-02 -4.210052e-03
#> 61  -94.59583 39.15833 -3.650850e-02 -8.487773e-03
#> 62  -84.43750 39.09583  1.114412e-02  2.586500e-03
#> 63  -95.60417 39.07083 -1.433011e-02  3.035717e-03
#> 64  -84.65417 39.00833  1.958403e-03  1.647046e-04
#> 65  -98.80000 38.90000 -6.817864e-04 -9.086475e-04
#> 66  -97.54167 38.81250 -5.194121e-03  2.290780e-03
#> 67  -92.17917 38.81250 -1.117192e-02 -2.040694e-03
#> 68  -90.37083 38.74583 -1.067519e-02 -7.441286e-03
#> 69  -90.62917 38.61250 -6.180026e-03 -5.002488e-03
#> 70  -82.62084 38.39583 -3.092829e-04 -5.426441e-03
#> 71  -81.62083 38.39583 -6.010979e-03 -6.830151e-03
#> 72  -85.72500 38.19167  1.758915e-02 -6.186110e-04
#> 73  -87.48333 38.00833 -8.948779e-04 -4.049524e-04
#> 74  -84.58333 38.00833  7.358828e-03 -5.666806e-03
#> 75  -81.09167 37.84167 -1.047206e-02 -4.912721e-03
#> 76  -99.96667 37.74167  9.031064e-04 -8.149519e-03
#> 77  -97.38333 37.66667  1.833404e-03  1.422064e-03
#> 78  -83.26667 37.53750 -2.549559e-02 -1.105528e-02
#> 79  -81.25417 37.35833 -1.861987e-02 -1.894913e-03
#> 80  -89.52083 37.24583  1.312128e-02  2.236495e-03
#> 81  -93.26250 37.20417  9.304700e-03  4.948397e-03
#> 82  -94.48750 37.14583  2.389799e-02  8.893025e-03
#> 83  -88.80000 37.06250  1.776967e-02 -2.754138e-03
#> 84  -86.46250 36.93333  8.379383e-03 -1.100632e-02
#> 85  -97.17083 36.72917  5.256938e-03 -6.024323e-03
#> 86  -82.37500 36.50000 -1.286473e-02  1.168896e-03
#> 87  -99.79584 36.32917  1.177465e-03 -3.084256e-03
#> 88  -92.97500 36.24167  1.761988e-03 -2.719467e-03
#> 89  -95.89584 36.18750  5.976349e-03  3.262285e-03
#> 90  -86.67500 36.11667  6.819524e-03  5.294241e-04
#> 91  -84.27083 36.03333  1.871886e-03  2.819848e-03
#> 92  -85.07500 35.93750  1.092730e-02 -1.454615e-03
#> 93  -83.96250 35.82917 -1.486301e-02 -4.992243e-04
#> 94  -81.35417 35.72083 -1.309813e-04  2.233861e-03
#> 95  -88.92500 35.59167 -3.100228e-02  3.582190e-03
#> 96  -82.47083 35.46667 -2.990012e-02 -1.832208e-03
#> 97  -97.60833 35.39167 -1.499625e-04 -3.139861e-02
#> 98  -94.35417 35.32083 -2.938308e-03  5.125860e-04
#> 99  -80.91250 35.20417 -1.280899e-02  1.536636e-03
#> 100 -89.93333 35.08750 -3.221705e-02  5.105109e-03
#> 101 -85.17500 35.02917 -5.885208e-03  3.512947e-03
#> 102 -99.15000 34.98750 -1.176916e-03 -3.480423e-02
#> 103 -95.76667 34.93333 -2.567433e-03 -3.245751e-03
#> 104 -82.20000 34.88750 -1.492402e-02  6.813688e-03
#> 105 -92.29167 34.82917  7.114803e-03  4.191473e-03
#> 106 -87.66667 34.76667 -1.935551e-02  5.964343e-03
#> 107 -92.23750 34.72917  1.563573e-03  2.458545e-03
#> 108 -86.57084 34.60000 -3.001345e-02  5.136889e-03
#> 109 -88.80000 34.30417 -9.514028e-03  8.964355e-03
#> 110 -98.57084 34.04583 -1.204707e-03 -3.642885e-02
#> 111 -83.26667 33.88333 -2.614492e-02  9.666363e-03
#> 112 -81.09167 33.88333 -8.079430e-03  4.084642e-03
#> 113 -84.37500 33.70833 -4.555034e-02  1.458509e-03
#> 114 -85.91666 33.59583 -8.303418e-02  3.582496e-03
#> 115 -86.85833 33.55417 -1.950921e-02 -2.101561e-03
#> 116 -90.21667 33.48333  3.194167e-03  3.045572e-03
#> 117 -82.01250 33.36667 -2.682311e-02  3.007373e-03
#> 118 -87.66667 33.18333 -2.186338e-02 -1.809297e-02
#> 119 -87.19583 32.97083 -6.233411e-02 -2.548822e-03
#> 120 -80.03333 32.97083 -5.099040e-02  4.400155e-03
#> 121 -96.87917 32.85833 -1.175829e-02 -6.583088e-02
#> 122 -97.01250 32.80417 -1.633446e-02 -5.821320e-02
#> 123 -83.63333 32.69583 -3.009380e-02  8.513808e-03
#> 124 -93.84167 32.57917 -3.495659e-02 -2.985687e-02
#> 125 -92.00417 32.50417 -2.240424e-03 -1.251782e-02
#> 126 -84.93333 32.50417 -3.697305e-02  8.467008e-03
#> 127 -99.54584 32.42917 -4.823208e-03 -1.011103e-01
#> 128 -88.60000 32.35417 -2.271334e-02 -8.166815e-03
#> 129 -90.08750 32.31667 -1.817048e-02 -2.247984e-02
#> 130 -86.38750 32.28333 -9.308236e-02 -4.054333e-03
#> 131 -98.06250 32.21667 -6.429458e-03 -1.135760e-01
#> 132 -81.18333 32.13334 -7.133143e-02  7.741188e-03
#> 
#> $locations_df
#>             lon      lat
#> 24    -95.37083 45.86666
#> 149   -94.12500 45.65000
#> 321   -84.79583 45.54167
#> 364   -98.42500 45.38333
#> 574   -83.52500 45.12500
#> 650   -89.62083 44.94583
#> 749   -93.20000 44.88750
#> 877   -91.48750 44.83750
#> 1029  -85.60417 44.72083
#> 1096  -95.06667 44.57917
#> 1252  -88.22917 44.47500
#> 1318  -98.23333 44.40417
#> 1512  -84.70417 44.27917
#> 1584  -92.60417 44.07500
#> 1711  -91.31250 43.85000
#> 1803  -96.68750 43.54167
#> 1877  -84.08750 43.54167
#> 1978  -86.14583 43.25417
#> 2056  -93.39167 43.15000
#> 2199  -89.33333 43.09583
#> 2356  -83.73333 43.00417
#> 2443  -87.88750 42.93750
#> 2577  -85.45834 42.87500
#> 2704  -84.58333 42.75000
#> 2775  -92.37917 42.57083
#> 2903  -90.81250 42.43333
#> 2994  -96.37917 42.35833
#> 3181  -84.49167 42.28750
#> 3312  -83.26667 42.23333
#> 3390  -89.11250 42.17917
#> 3569  -80.15417 42.08750
#> 3584  -97.38333 41.98333
#> 3633  -87.88750 41.98333
#> 3732  -91.73333 41.86250
#> 3880  -86.32500 41.72083
#> 4021  -83.82917 41.60833
#> 4078  -93.60000 41.52917
#> 4214  -90.51667 41.45833
#> 4392  -81.80000 41.40417
#> 4423  -96.07500 41.35833
#> 4543  -95.89584 41.30833
#> 4757  -80.62083 41.23333
#> 4797  -92.60417 41.12083
#> 4975  -83.67916 41.03333
#> 5077  -85.28750 40.99583
#> 5125  -98.32083 40.96250
#> 5346  -81.46250 40.91250
#> 5373  -96.68750 40.85833
#> 5576  -82.51667 40.78750
#> 5647  -89.74167 40.66250
#> 5829  -80.32083 40.41250
#> 5931  -82.90000 40.10417
#> 6057  -81.90417 39.95417
#> 6115  -91.09583 39.92917
#> 6279  -84.17916 39.87083
#> 6361  -89.74167 39.78333
#> 6499  -86.27083 39.69583
#> 6556  -97.77917 39.55417
#> 6774  -81.46250 39.39167
#> 6809  -94.77500 39.26667
#> 6929  -94.59583 39.15833
#> 7109  -84.43750 39.09583
#> 7163  -95.60417 39.07083
#> 7344  -84.65417 39.00833
#> 7383  -98.80000 38.90000
#> 7510  -97.54167 38.81250
#> 7538  -92.17917 38.81250
#> 7666  -90.37083 38.74583
#> 7783  -90.62917 38.61250
#> 7955  -82.62084 38.39583
#> 7963  -81.62083 38.39583
#> 8049  -85.72500 38.19167
#> 8157  -87.48333 38.00833
#> 8178  -84.58333 38.00833
#> 8325  -81.09167 37.84167
#> 8331  -99.96667 37.74167
#> 8463  -97.38333 37.66667
#> 8667  -83.26667 37.53750
#> 8799  -81.25417 37.35833
#> 8862  -89.52083 37.24583
#> 8959  -93.26250 37.20417
#> 9072  -94.48750 37.14583
#> 9223  -88.80000 37.06250
#> 9352  -86.46250 36.93333
#> 9416  -97.17083 36.72917
#> 9624  -82.37500 36.50000
#> 9641  -99.79584 36.32917
#> 9794  -92.97500 36.24167
#> 9898  -95.89584 36.18750
#> 10064 -86.67500 36.11667
#> 10205 -84.27083 36.03333
#> 10315 -85.07500 35.93750
#> 10446 -83.96250 35.82917
#> 10583 -81.35417 35.72083
#> 10650 -88.92500 35.59167
#> 10813 -82.47083 35.46667
#> 10841 -97.60833 35.39167
#> 10977 -94.35417 35.32083
#> 11182 -80.91250 35.20417
#> 11239 -89.93333 35.08750
#> 11385 -85.17500 35.02917
#> 11428 -99.15000 34.98750
#> 11565 -95.76667 34.93333
#> 11767 -82.20000 34.88750
#> 11820 -92.29167 34.82917
#> 11964 -87.66667 34.76667
#> 12059 -92.23750 34.72917
#> 12207 -86.57084 34.60000
#> 12317 -88.80000 34.30417
#> 12382 -98.57084 34.04583
#> 12594 -83.26667 33.88333
#> 12609 -81.09167 33.88333
#> 12703 -84.37500 33.70833
#> 12808 -85.91666 33.59583
#> 12919 -86.85833 33.55417
#> 13022 -90.21667 33.48333
#> 13196 -82.01250 33.36667
#> 13273 -87.66667 33.18333
#> 13394 -87.19583 32.97083
#> 13447 -80.03333 32.97083
#> 13464 -96.87917 32.85833
#> 13582 -97.01250 32.80417
#> 13782 -83.63333 32.69583
#> 13835 -93.84167 32.57917
#> 13965 -92.00417 32.50417
#> 14005 -84.93333 32.50417
#> 14045 -99.54584 32.42917
#> 14222 -88.60000 32.35417
#> 14332 -90.08750 32.31667
#> 14470 -86.38750 32.28333
#> 14528 -98.06250 32.21667
#> 14750 -81.18333 32.13334
#> 
#> $values_df
#>        [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8] [,9] [,10] [,11] [,12] [,13]
#>   [1,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#>   [2,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.04
#>   [3,] 0.00 0.00 0.00 0.00 0.00 0.04 0.01 0.01 0.00  0.09  0.01  0.00  0.05
#>   [4,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.00 0.00  0.00  0.05  0.00  0.00
#>   [5,] 0.00 0.00 0.00 0.05 0.00 0.01 0.04 0.00 0.00  0.05  0.00  0.00  0.01
#>   [6,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.11
#>   [7,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.00 0.00  0.00  0.00  0.00  0.04
#>   [8,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.01  0.00  0.00  0.12
#>   [9,] 0.00 0.00 0.00 0.00 0.00 0.00 0.07 0.08 0.14  0.08  0.08  0.00  0.00
#>  [10,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.01  0.00  0.00
#>  [11,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.02  0.00  0.00  0.06
#>  [12,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.38  0.02  0.00
#>  [13,] 0.00 0.00 0.00 0.00 0.00 0.01 0.04 0.00 0.01  0.04  0.02  0.00  0.00
#>  [14,] 0.00 0.00 0.01 0.00 0.00 0.00 0.03 0.00 0.00  0.00  0.00  0.00  0.06
#>  [15,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.01  0.00  0.00  0.00  0.03
#>  [16,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.25  0.03  0.00
#>  [17,] 0.00 0.00 0.00 0.00 0.00 0.00 0.02 0.02 0.00  0.01  0.00  0.00  0.01
#>  [18,] 0.00 0.00 0.00 0.00 0.00 0.00 0.04 0.02 0.04  0.11  0.00  0.00  0.01
#>  [19,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.01  0.00  0.01  0.00  0.03
#>  [20,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.01 0.01  0.13  0.00  0.00  0.02
#>  [21,] 0.00 0.00 0.00 0.00 0.00 0.00 0.08 0.00 0.00  0.01  0.03  0.00  0.00
#>  [22,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.00  0.05  0.00  0.00  0.00
#>  [23,] 0.00 0.00 0.00 0.00 0.00 0.00 0.02 0.01 0.00  0.07  0.01  0.00  0.00
#>  [24,] 0.00 0.00 0.00 0.00 0.00 0.00 0.09 0.04 0.00  0.02  0.00  0.00  0.00
#>  [25,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.01  0.02  0.00  0.00  0.00
#>  [26,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.02  0.32  0.00  0.00  0.00
#>  [27,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.01  0.00  0.00  0.00
#>  [28,] 0.00 0.00 0.00 0.00 0.00 0.00 0.19 0.03 0.01  0.00  0.01  0.00  0.00
#>  [29,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.01 0.00  0.00  0.00  0.00  0.01
#>  [30,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.04  0.00  0.00  0.00
#>  [31,] 0.00 0.00 0.00 0.08 0.00 0.00 0.03 0.03 0.00  0.00  0.17  0.00  0.14
#>  [32,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#>  [33,] 0.00 0.00 0.00 0.00 0.00 0.00 0.03 0.00 0.00  0.00  0.00  0.00  0.01
#>  [34,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.02  0.24  0.00  0.00  0.00
#>  [35,] 0.00 0.00 0.00 0.00 0.00 0.00 0.11 0.30 0.03  0.01  0.01  0.00  0.02
#>  [36,] 0.00 0.00 0.00 0.00 0.00 0.00 0.05 0.04 0.00  0.00  0.00  0.00  0.07
#>  [37,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.13  0.05  0.00  0.00  0.01
#>  [38,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.09  0.00  0.00  0.00
#>  [39,] 0.05 0.00 0.00 0.04 0.00 0.00 0.08 0.05 0.00  0.00  0.08  0.00  0.08
#>  [40,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.27  0.00  0.03  0.00
#>  [41,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.14  0.00  0.04  0.00
#>  [42,] 0.02 0.00 0.00 0.01 0.01 0.00 0.02 0.05 0.00  0.00  0.01  0.00  0.13
#>  [43,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.02  0.04  0.00  0.00  0.00
#>  [44,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.08 0.01  0.00  0.00  0.00  0.03
#>  [45,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.04 0.00  0.00  0.00  0.00  0.04
#>  [46,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.01  0.00  0.01  0.00
#>  [47,] 0.04 0.00 0.00 0.02 0.00 0.00 0.03 0.04 0.00  0.00  0.01  0.00  0.28
#>  [48,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.03  0.00  0.07  0.00
#>  [49,] 0.00 0.00 0.00 0.00 0.00 0.00 0.03 0.11 0.00  0.00  0.00  0.00  0.07
#>  [50,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.03  0.00  0.05  0.00
#>  [51,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.05 0.00  0.00  0.00  0.00  0.34
#>  [52,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.05 0.00  0.00  0.00  0.00  0.30
#>  [53,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.02 0.00  0.00  0.00  0.00  0.23
#>  [54,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.02  0.00  0.06  0.00
#>  [55,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.02 0.00  0.00  0.00  0.00  0.14
#>  [56,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.02 0.00  0.00  0.00  0.03  0.00
#>  [57,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.00  0.00  0.00  0.02  0.07
#>  [58,] 0.00 0.00 0.04 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.45  0.05  0.00
#>  [59,] 0.01 0.00 0.00 0.00 0.00 0.00 0.02 0.01 0.01  0.00  0.00  0.04  0.52
#>  [60,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.03  0.17  0.00
#>  [61,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.15  0.00
#>  [62,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.05  0.16
#>  [63,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.15  0.08  0.00
#>  [64,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.00  0.00  0.00  0.10  0.25
#>  [65,] 0.00 0.00 0.13 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.41  0.01  0.01
#>  [66,] 0.00 0.00 0.09 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.27  0.03  0.00
#>  [67,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.28  0.00  0.41  0.00
#>  [68,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.37  0.00
#>  [69,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.44  0.00
#>  [70,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.03 0.00  0.00  0.00  0.08  0.64
#>  [71,] 0.01 0.00 0.00 0.00 0.00 0.00 0.01 0.04 0.00  0.00  0.00  0.01  0.36
#>  [72,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.28  0.29
#>  [73,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.54  0.00
#>  [74,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.02 0.00  0.00  0.00  0.33  0.55
#>  [75,] 0.05 0.00 0.00 0.00 0.00 0.00 0.00 0.03 0.01  0.00  0.00  0.00  0.16
#>  [76,] 0.00 0.03 0.14 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.03  0.00  0.00
#>  [77,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.17  0.00  0.00
#>  [78,] 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.04 0.00  0.00  0.00  0.39  0.54
#>  [79,] 0.05 0.00 0.00 0.00 0.00 0.00 0.00 0.01 0.00  0.00  0.00  0.00  0.06
#>  [80,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.03  0.56  0.30
#>  [81,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.20  0.21  0.43  0.01
#>  [82,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.21  0.84  0.27  0.01
#>  [83,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.07  0.71  0.41
#>  [84,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.85  0.62
#>  [85,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.01  0.22  0.00  0.05
#>  [86,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.10
#>  [87,] 0.00 0.00 0.19 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#>  [88,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.61  0.01  0.50  0.02
#>  [89,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.02  0.41  0.54  0.05  0.00
#>  [90,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.12  0.26
#>  [91,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.01  0.38
#>  [92,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.14  0.38
#>  [93,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.25
#>  [94,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.14
#>  [95,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.01  0.05  0.07
#>  [96,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.01
#>  [97,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.01  0.05  0.38  0.00  0.00
#>  [98,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.04  0.93  0.00  0.00
#>  [99,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.08  0.00  0.07
#> [100,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.08  0.02  0.01
#> [101,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.35
#> [102,] 0.00 0.12 0.07 0.09 0.00 0.00 0.00 0.00 0.00  0.00  0.01  0.01  0.00
#> [103,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.43  0.00  0.00
#> [104,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.01
#> [105,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.95  0.16  0.00
#> [106,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.05  0.01
#> [107,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.82  0.15  0.00
#> [108,] 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.05  0.04
#> [109,] 0.00 0.00 0.00 0.01 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.07  0.00
#> [110,] 0.00 0.00 0.50 0.00 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#> [111,] 0.00 0.00 0.00 0.00 0.10 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.09
#> [112,] 0.00 0.00 0.00 0.00 0.12 0.00 0.00 0.00 0.00  0.00  0.01  0.00  0.01
#> [113,] 0.00 0.00 0.00 0.00 0.18 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.05
#> [114,] 0.00 0.00 0.00 0.00 0.27 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.11
#> [115,] 0.00 0.00 0.00 0.00 0.31 0.00 0.00 0.00 0.00  0.00  0.00  0.01  0.09
#> [116,] 0.00 0.00 0.00 0.05 0.03 0.00 0.00 0.00 0.00  0.00  0.00  0.71  0.00
#> [117,] 0.00 0.00 0.00 0.00 0.16 0.19 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#> [118,] 0.00 0.00 0.00 0.00 0.59 0.04 0.00 0.00 0.00  0.00  0.00  0.13  0.11
#> [119,] 0.00 0.00 0.00 0.00 0.86 0.03 0.00 0.00 0.00  0.00  0.00  0.08  0.19
#> [120,] 0.00 0.00 0.00 0.00 0.10 0.25 0.00 0.00 0.00  0.02  0.30  0.00  0.00
#> [121,] 0.00 0.00 0.03 0.05 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.12  0.00
#> [122,] 0.00 0.01 0.00 0.10 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.25  0.00
#> [123,] 0.00 0.00 0.00 0.00 0.65 0.35 0.00 0.00 0.00  0.00  0.00  0.00  0.03
#> [124,] 0.00 0.00 0.00 0.92 0.95 0.00 0.00 0.00 0.00  0.00  0.05  1.10  0.00
#> [125,] 0.00 0.00 0.00 0.50 0.58 0.00 0.00 0.00 0.00  0.00  0.00  0.97  0.00
#> [126,] 0.00 0.00 0.00 0.00 0.20 0.16 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#> [127,] 0.00 0.16 0.95 0.27 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#> [128,] 0.00 0.00 0.00 0.01 1.25 0.02 0.00 0.00 0.00  0.00  0.00  0.27  0.66
#> [129,] 0.00 0.00 0.00 0.05 0.92 0.00 0.00 0.00 0.00  0.00  0.00  1.22  0.70
#> [130,] 0.00 0.00 0.00 0.00 1.21 0.13 0.00 0.00 0.00  0.00  0.00  0.00  0.01
#> [131,] 0.00 0.43 0.93 0.56 0.00 0.00 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#> [132,] 0.00 0.00 0.00 0.00 0.00 0.74 0.00 0.00 0.00  0.00  0.00  0.00  0.00
#>        [,14] [,15] [,16] [,17] [,18] [,19] [,20] [,21] [,22] [,23] [,24] [,25]
#>   [1,]  0.00  0.08  0.00  0.00  0.10  0.03  0.11  0.00  0.00  0.00  0.04  0.00
#>   [2,]  0.00  0.00  0.00  0.00  0.00  0.03  0.04  0.00  0.00  0.00  0.16  0.00
#>   [3,]  0.00  0.12  0.00  0.00  0.10  0.06  0.30  0.00  0.02  0.02  0.07  0.29
#>   [4,]  0.01  0.04  0.01  0.15  0.13  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#>   [5,]  0.00  0.57  0.00  0.00  0.24  0.08  0.13  0.00  0.03  0.00  0.00  0.27
#>   [6,]  0.00  0.00  0.00  0.00  0.19  0.03  0.02  0.00  0.00  0.03  0.42  0.02
#>   [7,]  0.00  0.00  0.00  0.00  0.15  0.01  0.00  0.02  0.00  0.01  0.33  0.00
#>   [8,]  0.00  0.00  0.00  0.00  0.09  0.00  0.00  0.00  0.00  0.06  0.36  0.00
#>   [9,]  0.00  0.30  0.00  0.00  0.20  0.03  0.19  0.00  0.00  0.00  0.05  0.16
#>  [10,]  0.00  0.00  0.00  0.03  0.23  0.00  0.00  0.04  0.00  0.03  0.10  0.00
#>  [11,]  0.00  0.09  0.00  0.00  0.14  0.02  0.02  0.00  0.00  0.00  0.17  0.01
#>  [12,]  0.06  0.01  0.00  0.31  0.25  0.00  0.00  0.14  0.00  0.00  0.00  0.00
#>  [13,]  0.00  0.31  0.00  0.00  0.19  0.18  0.14  0.01  0.00  0.00  0.09  0.20
#>  [14,]  0.01  0.05  0.00  0.00  0.25  0.01  0.00  0.01  0.00  0.00  0.12  0.00
#>  [15,]  0.00  0.15  0.00  0.00  0.39  0.09  0.00  0.00  0.00  0.00  0.14  0.00
#>  [16,]  0.23  0.01  0.07  0.61  0.18  0.00  0.00  0.01  0.00  0.25  0.12  0.00
#>  [17,]  0.00  0.48  0.00  0.00  0.21  0.02  0.04  0.00  0.00  0.00  0.19  0.17
#>  [18,]  0.00  0.43  0.00  0.00  0.25  0.00  0.01  0.00  0.00  0.00  0.26  0.07
#>  [19,]  0.11  0.16  0.01  0.06  0.62  0.00  0.00  0.02  0.00  0.00  0.12  0.00
#>  [20,]  0.01  0.48  0.00  0.00  0.34  0.02  0.00  0.00  0.00  0.00  0.31  0.00
#>  [21,]  0.00  0.55  0.02  0.00  0.12  0.30  0.04  0.00  0.05  0.02  0.00  0.22
#>  [22,]  0.07  0.38  0.00  0.00  0.70  0.06  0.00  0.00  0.00  0.00  0.24  0.00
#>  [23,]  0.00  0.43  0.02  0.00  0.10  0.00  0.00  0.00  0.04  0.00  0.16  0.05
#>  [24,]  0.00  0.55  0.02  0.00  0.19  0.24  0.01  0.00  0.00  0.00  0.05  0.10
#>  [25,]  0.37  0.27  0.00  0.35  0.21  0.02  0.00  0.00  0.00  0.00  0.22  0.00
#>  [26,]  0.20  0.33  0.00  0.04  0.65  0.10  0.00  0.00  0.00  0.00  0.21  0.00
#>  [27,]  0.36  0.01  0.00  0.24  0.04  0.01  0.00  0.00  0.00  0.06  0.04  0.00
#>  [28,]  0.00  0.41  0.02  0.00  0.05  0.07  0.01  0.00  0.03  0.00  0.00  0.12
#>  [29,]  0.00  0.77  0.09  0.00  0.21  0.06  0.04  0.00  0.09  0.00  0.00  0.14
#>  [30,]  0.07  0.27  0.00  0.06  0.40  0.02  0.00  0.00  0.00  0.00  0.33  0.00
#>  [31,]  0.02  0.43  0.19  0.00  0.09  0.27  0.03  0.00  0.06  0.00  0.02  0.01
#>  [32,]  0.41  0.01  0.10  0.51  0.00  0.00  0.00  0.00  0.00  0.41  0.04  0.00
#>  [33,]  0.21  0.40  0.00  0.01  0.30  0.05  0.00  0.00  0.00  0.00  0.36  0.00
#>  [34,]  0.50  0.03  0.00  0.38  0.29  0.00  0.00  0.00  0.00  0.00  0.14  0.00
#>  [35,]  0.10  0.50  0.00  0.00  0.20  0.10  0.00  0.00  0.10  0.00  0.06  0.12
#>  [36,]  0.02  0.78  0.04  0.00  0.33  0.05  0.00  0.00  0.13  0.00  0.00  0.09
#>  [37,]  0.96  0.17  0.00  0.52  0.13  0.01  0.00  0.00  0.00  0.11  0.03  0.00
#>  [38,]  0.62  0.14  0.00  0.28  0.25  0.02  0.00  0.00  0.00  0.05  0.46  0.00
#>  [39,]  0.01  0.60  0.07  0.00  0.43  0.30  0.04  0.00  0.05  0.00  0.00  0.00
#>  [40,]  0.30  0.02  0.00  0.41  0.13  0.00  0.00  0.00  0.00  0.01  0.04  0.00
#>  [41,]  0.28  0.06  0.00  0.46  0.13  0.02  0.00  0.00  0.00  0.00  0.05  0.00
#>  [42,]  0.01  0.42  0.02  0.00  0.08  0.15  0.01  0.00  0.08  0.00  0.04  0.00
#>  [43,]  0.49  0.03  0.00  0.60  0.15  0.02  0.00  0.00  0.00  0.42  0.10  0.00
#>  [44,]  0.02  0.32  0.00  0.00  0.22  0.02  0.01  0.00  0.01  0.00  0.00  0.00
#>  [45,]  0.09  0.81  0.00  0.00  0.15  0.15  0.00  0.00  0.00  0.00  0.00  0.11
#>  [46,]  0.50  0.00  0.07  0.58  0.01  0.00  0.00  0.00  0.00  0.10  0.01  0.00
#>  [47,]  0.03  0.38  0.08  0.00  0.06  0.25  0.05  0.00  0.07  0.01  0.06  0.01
#>  [48,]  0.31  0.00  0.02  0.22  0.23  0.00  0.00  0.00  0.00  0.00  0.05  0.00
#>  [49,]  0.00  0.66  0.02  0.00  0.25  0.29  0.02  0.00  0.01  0.00  0.00  0.00
#>  [50,]  0.83  0.01  0.00  0.19  0.08  0.01  0.00  0.02  0.01  0.00  0.32  0.00
#>  [51,]  0.00  0.51  0.04  0.00  0.06  0.05  0.05  0.00  0.00  0.01  0.26  0.00
#>  [52,]  0.00  0.26  0.00  0.00  0.08  0.07  0.02  0.00  0.00  0.02  0.01  0.00
#>  [53,]  0.00  0.49  0.00  0.00  0.03  0.07  0.04  0.00  0.00  0.11  0.04  0.00
#>  [54,]  0.35  0.08  0.00  0.61  0.04  0.01  0.00  0.00  0.00  0.00  0.30  0.00
#>  [55,]  0.11  0.19  0.00  0.00  0.33  0.15  0.00  0.00  0.00  0.01  0.00  0.00
#>  [56,]  0.58  0.57  0.00  0.19  0.02  0.01  0.00  0.00  0.00  0.00  0.16  0.00
#>  [57,]  0.21  0.43  0.00  0.11  0.27  0.02  0.00  0.00  0.00  0.00  0.00  0.01
#>  [58,]  0.12  0.00  0.05  0.08  0.03  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#>  [59,]  0.01  0.17  0.03  0.00  0.06  0.04  0.06  0.00  0.00  0.22  0.06  0.09
#>  [60,]  0.50  0.00  0.00  1.03  0.01  0.00  0.00  0.00  0.00  0.10  0.17  0.00
#>  [61,]  0.64  0.00  0.00  0.87  0.00  0.00  0.00  0.00  0.00  0.03  0.27  0.00
#>  [62,]  0.02  0.54  0.00  0.00  0.06  0.06  0.00  0.00  0.00  0.06  0.00  0.00
#>  [63,]  0.47  0.00  0.02  0.41  0.02  0.00  0.00  0.00  0.00  0.01  0.02  0.00
#>  [64,]  0.01  0.26  0.02  0.02  0.14  0.06  0.01  0.00  0.00  0.02  0.00  0.00
#>  [65,]  0.08  0.00  0.07  0.09  0.00  0.00  0.00  0.00  0.00  0.00  0.02  0.00
#>  [66,]  0.32  0.00  0.02  0.22  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00
#>  [67,]  0.71  0.00  0.00  0.43  0.01  0.00  0.00  0.00  0.00  0.00  0.12  0.00
#>  [68,]  0.79  0.00  0.00  0.45  0.00  0.00  0.00  0.00  0.00  0.06  0.22  0.00
#>  [69,]  0.82  0.04  0.00  0.43  0.00  0.00  0.00  0.00  0.00  0.13  0.17  0.00
#>  [70,]  0.01  0.17  0.00  0.03  0.02  0.02  0.00  0.00  0.00  0.50  0.03  0.43
#>  [71,]  0.00  0.35  0.00  0.01  0.02  0.01  0.02  0.00  0.00  0.56  0.10  0.88
#>  [72,]  0.15  0.06  0.00  0.03  0.01  0.01  0.00  0.00  0.00  0.71  0.00  0.00
#>  [73,]  0.39  0.00  0.00  0.27  0.05  0.00  0.00  0.00  0.00  0.26  0.00  0.00
#>  [74,]  0.06  0.07  0.00  0.07  0.00  0.02  0.00  0.00  0.00  0.43  0.02  0.02
#>  [75,]  0.00  0.34  0.10  0.11  0.03  0.00  0.00  0.00  0.00  0.32  0.12  0.67
#>  [76,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.02  0.13  0.00
#>  [77,]  0.14  0.00  0.02  0.04  0.00  0.00  0.00  0.00  0.00  0.05  0.01  0.00
#>  [78,]  0.01  0.25  0.00  0.01  0.01  0.01  0.00  0.00  0.00  0.16  0.01  1.31
#>  [79,]  0.00  0.30  0.00  0.04  0.06  0.05  0.00  0.00  0.00  0.56  0.03  1.31
#>  [80,]  0.91  0.00  0.00  0.37  0.00  0.00  0.00  0.00  0.00  0.81  0.00  0.00
#>  [81,]  1.15  0.00  0.00  0.17  0.00  0.00  0.00  0.00  0.00  0.00  0.03  0.00
#>  [82,]  1.54  0.00  0.00  0.02  0.00  0.00  0.00  0.00  0.00  0.00  0.06  0.00
#>  [83,]  0.64  0.00  0.00  0.16  0.02  0.00  0.00  0.00  0.00  0.67  0.00  0.00
#>  [84,]  0.15  0.02  0.00  0.07  0.02  0.00  0.00  0.00  0.00  0.72  0.00  0.35
#>  [85,]  0.20  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.11  0.16  0.00
#>  [86,]  0.00  0.21  0.00  0.05  0.04  0.00  0.00  0.00  0.00  0.92  0.00  1.34
#>  [87,]  0.00  0.00  0.07  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.02  0.00
#>  [88,]  0.13  0.00  0.00  0.04  0.00  0.00  0.00  0.00  0.00  0.16  0.04  0.20
#>  [89,]  0.17  0.00  0.01  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.03  0.00
#>  [90,]  0.18  0.17  0.00  0.06  0.07  0.00  0.00  0.00  0.00  0.97  0.00  0.63
#>  [91,]  0.14  0.63  0.01  0.01  0.09  0.00  0.00  0.00  0.00  0.52  0.00  0.85
#>  [92,]  0.15  0.29  0.00  0.01  0.02  0.00  0.00  0.00  0.00  0.85  0.05  0.58
#>  [93,]  0.05  0.50  0.00  0.06  0.04  0.00  0.00  0.00  0.00  0.61  0.00  1.37
#>  [94,]  0.00  0.92  0.00  0.06  0.07  0.00  0.00  0.00  0.00  0.73  0.00  1.21
#>  [95,]  0.40  0.14  0.00  0.74  0.00  0.00  0.00  0.00  0.03  0.67  0.00  0.62
#>  [96,]  0.00  0.49  0.00  0.18  0.00  0.01  0.00  0.00  0.00  0.79  0.00  1.89
#>  [97,]  0.02  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.17  0.00  0.65  0.00
#>  [98,]  0.00  0.00  0.01  0.04  0.00  0.00  0.00  0.00  0.17  0.03  0.00  0.25
#>  [99,]  0.00  0.85  0.01  0.21  0.08  0.04  0.00  0.00  0.00  0.89  0.05  1.51
#> [100,]  0.74  0.00  0.00  0.65  0.00  0.00  0.00  0.00  0.03  0.15  0.00  0.45
#> [101,]  0.30  0.75  0.00  0.40  0.02  0.00  0.00  0.00  0.00  1.29  0.00  1.26
#> [102,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.05  0.00  0.00  0.68  0.00
#> [103,]  0.00  0.00  0.02  0.00  0.00  0.00  0.00  0.00  0.12  0.00  0.06  0.18
#> [104,]  0.00  1.12  0.40  0.30  0.13  0.02  0.00  0.00  0.00  1.51  0.00  2.19
#> [105,]  0.70  0.00  0.00  0.04  0.00  0.00  0.00  0.00  0.15  0.05  0.00  0.25
#> [106,]  0.35  0.15  0.00  0.51  0.24  0.00  0.00  0.00  0.10  0.92  0.00  0.64
#> [107,]  0.52  0.00  0.00  0.06  0.00  0.00  0.00  0.00  0.14  0.05  0.00  0.31
#> [108,]  0.45  0.30  0.00  0.69  0.16  0.00  0.00  0.00  0.11  0.69  0.00  0.72
#> [109,]  1.17  0.00  0.00  0.51  0.01  0.00  0.00  0.00  0.02  0.77  0.00  0.59
#> [110,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.60  0.05
#> [111,]  0.03  1.12  0.00  0.77  0.10  0.00  0.00  0.00  0.00  0.67  0.00  0.96
#> [112,]  0.00  0.47  0.00  0.38  0.13  0.01  0.00  0.00  0.00  1.60  0.00  1.25
#> [113,]  0.14  0.86  0.00  0.78  0.01  0.05  0.00  0.00  0.00  0.42  0.00  1.35
#> [114,]  0.59  0.99  0.00  0.87  0.55  0.00  0.00  0.00  0.20  0.28  0.00  2.36
#> [115,]  0.14  1.03  0.00  0.48  0.01  0.00  0.00  0.00  0.12  0.53  0.06  1.18
#> [116,]  1.60  0.00  0.00  0.14  0.00  0.00  0.00  0.00  0.00  0.10  0.00  0.63
#> [117,]  0.00  0.70  0.00  0.63  0.04  0.00  0.00  0.00  0.02  1.30  0.00  1.42
#> [118,]  0.28  1.27  0.00  0.67  0.00  0.00  0.00  0.00  0.20  0.43  0.39  1.08
#> [119,]  0.09  1.43  0.00  1.32  0.00  0.00  0.00  0.00  0.29  0.01  0.02  1.09
#> [120,]  0.00  0.01  0.00  0.99  0.11  0.03  0.00  0.00  0.00  0.27  0.00  0.15
#> [121,]  0.02  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.73  0.00  1.23  0.35
#> [122,]  0.01  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.41  0.00  1.05  0.57
#> [123,]  0.03  1.12  0.00  1.13  0.14  0.00  0.00  0.00  0.04  0.69  0.00  0.31
#> [124,]  0.03  0.00  0.18  0.00  0.00  0.00  0.00  0.00  1.05  0.05  0.04  2.04
#> [125,]  1.48  0.00  0.00  0.03  0.00  0.00  0.00  0.00  0.08  0.04  0.11  1.15
#> [126,]  0.11  1.62  0.00  1.32  0.00  0.00  0.00  0.00  0.54  0.47  0.00  0.60
#> [127,]  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.25  0.00  0.00  1.77  0.20
#> [128,]  0.02  1.71  0.00  0.75  0.00  0.00  0.00  0.00  0.23  0.07  0.00  0.98
#> [129,]  0.05  0.06  0.00  0.31  0.00  0.00  0.00  0.00  0.05  0.00  0.00  0.47
#> [130,]  0.10  1.80  0.00  1.53  0.00  0.00  0.00  0.00  0.80  0.06  0.03  2.26
#> [131,]  0.02  0.00  0.00  0.00  0.00  0.00  0.00  0.00  0.02  0.00  1.98  0.31
#> [132,]  0.01  0.00  0.00  1.17  0.25  0.00  0.00  0.00  0.00  0.03  0.00  0.04
#>        [,26] [,27] [,28]
#>   [1,]  0.00  0.00  0.00
#>   [2,]  0.00  0.00  0.00
#>   [3,]  0.00  0.00  0.25
#>   [4,]  0.07  0.00  0.00
#>   [5,]  0.03  0.00  0.19
#>   [6,]  0.00  0.00  0.00
#>   [7,]  0.00  0.00  0.00
#>   [8,]  0.00  0.00  0.00
#>   [9,]  0.00  0.00  0.11
#>  [10,]  0.00  0.00  0.00
#>  [11,]  0.00  0.00  0.02
#>  [12,]  0.01  0.00  0.00
#>  [13,]  0.00  0.00  0.11
#>  [14,]  0.00  0.00  0.00
#>  [15,]  0.00  0.00  0.00
#>  [16,]  0.00  0.00  0.00
#>  [17,]  0.00  0.00  0.20
#>  [18,]  0.00  0.00  0.09
#>  [19,]  0.00  0.00  0.00
#>  [20,]  0.00  0.00  0.00
#>  [21,]  0.00  0.00  0.15
#>  [22,]  0.00  0.00  0.03
#>  [23,]  0.00  0.00  0.15
#>  [24,]  0.00  0.00  0.05
#>  [25,]  0.00  0.00  0.00
#>  [26,]  0.00  0.00  0.00
#>  [27,]  0.00  0.00  0.00
#>  [28,]  0.02  0.00  0.00
#>  [29,]  0.00  0.00  0.11
#>  [30,]  0.00  0.00  0.00
#>  [31,]  0.00  0.00  0.34
#>  [32,]  0.00  0.00  0.00
#>  [33,]  0.00  0.00  0.02
#>  [34,]  0.00  0.00  0.00
#>  [35,]  0.00  0.00  0.05
#>  [36,]  0.00  0.00  0.08
#>  [37,]  0.00  0.00  0.00
#>  [38,]  0.00  0.00  0.00
#>  [39,]  0.00  0.00  0.76
#>  [40,]  0.00  0.00  0.00
#>  [41,]  0.00  0.00  0.00
#>  [42,]  0.00  0.00  0.58
#>  [43,]  0.00  0.00  0.00
#>  [44,]  0.00  0.00  0.15
#>  [45,]  0.00  0.00  0.00
#>  [46,]  0.07  0.00  0.00
#>  [47,]  0.00  0.00  0.45
#>  [48,]  0.00  0.00  0.00
#>  [49,]  0.00  0.00  0.62
#>  [50,]  0.00  0.00  0.00
#>  [51,]  0.00  0.00  0.35
#>  [52,]  0.00  0.00  0.04
#>  [53,]  0.00  0.00  0.14
#>  [54,]  0.00  0.00  0.00
#>  [55,]  0.00  0.00  0.00
#>  [56,]  0.00  0.00  0.00
#>  [57,]  0.00  0.00  0.00
#>  [58,]  0.06  0.00  0.00
#>  [59,]  0.05  0.00  0.20
#>  [60,]  0.00  0.00  0.00
#>  [61,]  0.00  0.00  0.00
#>  [62,]  0.00  0.00  0.18
#>  [63,]  0.00  0.00  0.00
#>  [64,]  0.00  0.00  0.03
#>  [65,]  0.02  0.00  0.00
#>  [66,]  0.05  0.02  0.00
#>  [67,]  0.00  0.00  0.00
#>  [68,]  0.00  0.00  0.00
#>  [69,]  0.00  0.00  0.00
#>  [70,]  0.15  0.00  0.19
#>  [71,]  0.05  0.00  0.29
#>  [72,]  0.00  0.00  0.20
#>  [73,]  0.00  0.00  0.00
#>  [74,]  0.04  0.00  0.21
#>  [75,]  0.07  0.00  0.28
#>  [76,]  0.00  0.00  0.00
#>  [77,]  0.00  0.04  0.00
#>  [78,]  0.20  0.00  0.17
#>  [79,]  0.21  0.00  0.12
#>  [80,]  0.00  0.00  0.00
#>  [81,]  0.00  0.00  0.00
#>  [82,]  0.00  0.00  0.00
#>  [83,]  0.00  0.00  0.00
#>  [84,]  0.02  0.00  0.06
#>  [85,]  0.00  0.00  0.00
#>  [86,]  0.59  0.00  0.09
#>  [87,]  0.00  0.00  0.00
#>  [88,]  0.00  0.00  0.00
#>  [89,]  0.00  0.09  0.00
#>  [90,]  0.14  0.00  0.00
#>  [91,]  0.59  0.00  0.02
#>  [92,]  0.44  0.00  0.09
#>  [93,]  0.47  0.00  0.01
#>  [94,]  0.25  0.00  0.00
#>  [95,]  0.10  0.00  0.00
#>  [96,]  0.28  0.00  0.00
#>  [97,]  0.00  0.00  0.00
#>  [98,]  0.00  0.01  0.00
#>  [99,]  0.52  0.00  0.00
#> [100,]  0.05  0.00  0.00
#> [101,]  0.26  0.00  0.00
#> [102,]  0.00  0.00  0.00
#> [103,]  0.00  0.00  0.00
#> [104,]  0.44  0.00  0.00
#> [105,]  0.00  0.00  0.00
#> [106,]  0.08  0.00  0.00
#> [107,]  0.00  0.00  0.00
#> [108,]  0.00  0.00  0.00
#> [109,]  0.02  0.00  0.00
#> [110,]  0.00  0.00  0.00
#> [111,]  0.91  0.00  0.00
#> [112,]  0.18  0.00  0.00
#> [113,]  0.10  0.00  0.00
#> [114,]  0.10  0.00  0.00
#> [115,]  0.00  0.00  0.00
#> [116,]  0.02  0.00  0.00
#> [117,]  0.04  0.00  0.00
#> [118,]  0.01  0.00  0.00
#> [119,]  0.00  0.00  0.00
#> [120,]  0.00  0.00  0.00
#> [121,]  0.00  0.00  0.00
#> [122,]  0.00  0.00  0.00
#> [123,]  0.00  0.00  0.00
#> [124,]  0.00  0.00  0.00
#> [125,]  0.00  0.00  0.00
#> [126,]  0.05  0.00  0.00
#> [127,]  0.00  0.00  0.00
#> [128,]  0.01  0.00  0.00
#> [129,]  0.00  0.00  0.00
#> [130,]  0.29  0.00  0.00
#> [131,]  0.00  0.00  0.00
#> [132,]  0.00  0.00  0.00
#> 
#> $cancor_obj
#> $cancor_obj$cor
#>  [1] 0.99999894 0.99846961 0.99491661 0.96047318 0.87195085 0.62666209
#>  [7] 0.32087986 0.19055342 0.09236795 0.05843937
#> 
#> $cancor_obj$xcoef
#>              [,1]        [,2]       [,3]        [,4]        [,5]         [,6]
#>  [1,] -0.15608897  0.53888078 -1.6176275  0.80222264 -1.37602273 -0.002113482
#>  [2,] -0.03600625 -0.03904464  0.3053663 -0.19562814  0.15023799  0.071826557
#>  [3,]  0.07560267  0.39258933 -1.0018040  0.41015069 -1.10724604 -0.027864476
#>  [4,] -0.04457063 -0.07156318 -0.4140306  0.19314062  0.15786860  0.100853062
#>  [5,] -0.01968479  0.03056024 -0.3306721  0.33222177  0.09825222 -0.248736928
#>  [6,]  0.11089105 -0.15390387  0.1216803 -0.02081427  0.27063538  0.103015429
#>  [7,] -0.02845212  0.06859196 -0.5905054  0.28609275 -0.28108343 -0.117240435
#>  [8,]  0.03479782  0.08019902  0.1722797 -0.27184084  0.16786550  0.065993380
#>  [9,]  0.06507451  0.11450232  0.1921201  0.02172699  0.03517279  0.029553428
#> [10,]  0.04568448  0.01052740  0.2763775 -0.23897723  0.05288691 -0.041216872
#>              [,7]         [,8]        [,9]         [,10]
#>  [1,]  0.42558763  0.964063701 -0.08986837  0.1942398073
#>  [2,] -0.06747275 -0.233143473 -0.09430817  0.1080615488
#>  [3,]  0.27276471  0.519334104 -0.20684227  0.0877549020
#>  [4,] -0.47950477  0.007542561  0.35378883 -0.0344135028
#>  [5,] -0.09703785  0.629631246  0.30592435  0.2475615927
#>  [6,] -0.05179212  0.006712656  0.07581296  0.0449991528
#>  [7,]  0.05576213  0.510298126  0.36290385  0.1996128840
#>  [8,]  0.01687358 -0.234635746 -0.12984387 -0.0956629609
#>  [9,]  0.06049424 -0.127711235  0.01233449 -0.0006896339
#> [10,]  0.19033093 -0.545402510 -0.38872651 -0.0576475262
#> 
#> $cancor_obj$ycoef
#>                [,1]        [,2]         [,3]         [,4]         [,5]
#>  [1,] -0.0004938218 -0.13590647 -0.039397306 -0.021777069  0.013178394
#>  [2,]  0.0060309775  0.02216956  0.004041259  0.006939533  0.018851100
#>  [3,] -0.0278779349 -0.04673222 -0.051452538  0.023178481 -0.027074582
#>  [4,] -0.1834819266 -0.02408094 -0.005569132  0.013579880  0.034287518
#>  [5,]  0.0478767251 -0.05244448 -0.017738948  0.112529013  0.171005134
#>  [6,]  0.0126459341  0.03282611 -0.017133610 -0.038438155 -0.001141411
#>  [7,]  0.0556230523 -0.05651993 -0.087717245 -0.039234779 -0.089694839
#>  [8,] -0.0236627957  0.02840493 -0.074427800  0.123122870 -0.075142373
#>  [9,]  0.0029106729 -0.04710897  0.136041316  0.116701496 -0.048556857
#> [10,]  0.0085012992  0.07390252 -0.067653000  0.061534455  0.058091028
#>              [,6]         [,7]         [,8]         [,9]        [,10]
#>  [1,] -0.08155481  0.112655513  0.002897841  0.005083706 -0.030900286
#>  [2,] -0.09021655 -0.004013168 -0.052100317  0.142014037  0.080404129
#>  [3,] -0.02185555 -0.066954323  0.128553973 -0.013009225  0.105270028
#>  [4,]  0.05188661 -0.013883631 -0.017994454  0.041350917 -0.030137279
#>  [5,]  0.02939708 -0.047450368 -0.032013564 -0.011599593 -0.034381090
#>  [6,] -0.07852340 -0.047632472  0.098413437  0.068991490 -0.125816323
#>  [7,]  0.08883402 -0.049296218 -0.025407653  0.080929158 -0.013334479
#>  [8,] -0.07138110 -0.001444693 -0.053986548 -0.033834697 -0.042522312
#>  [9,]  0.05800943  0.020733593  0.039944481  0.045376734 -0.017535745
#> [10,]  0.07248174  0.120921396  0.059934378  0.037521239  0.008182189
#> 
#> $cancor_obj$xcenter
#>  [1] -0.180364318  0.098503822 -0.217412717 -0.118650533  0.143353058
#>  [6]  0.038588864 -0.009881163 -0.009983706  0.005155323  0.009690939
#> 
#> $cancor_obj$ycenter
#>  [1]  0.13468471  0.08955597  0.01164330 -0.05936525 -0.11165454  0.14525823
#>  [7]  0.02239443  0.05745972 -0.17368192 -0.07229331
#> 
#> 
#> $call
#> cancor_eof.data.frame(x = locations_df2, lag = lag, n_eof = n_eof, 
#>     values_df = values_df2)
#> 
#> attr(,"class")
#> [1] "cancoreof"