Computes empirical spatial covariance using a stars object

Computes empirical spatial covariance by removing trends and examining residuals. It can compute lag-0 or log-1 empirical covariance either by latitude or longitude. You can split up the spatial domain by latitude or longitude and plot the covariance for each longitudinal/latitudinal strips.

Usage

# S3 method for stars
emp_spatial_cov(
  x,
  lat_or_lon_strips = "lon",
  quadratic_time = FALSE,
  quadratic_space = FALSE,
  num_strips = 1,
  lag = 0,
  ...
)

Arguments

x

A stars object or a dataframe. Arguments differ according to the input type.

lat_or_lon_strips

Takes the values lat or lon. The value lat produces latitudinal strips, i.e., covariance plots over longitude for different latitudinal strips. The value lon produces longitudinal strips, i.e., covariance plots over latitude for different longitudinal strips.

quadratic_time

If TRUE a linear model with quadratic time is fitted and residuals computed. If FALSE the model is fitted with linear space and time coefficients.

quadratic_space

If TRUE a linear model with quadratic space is fitted and residuals computed. If FALSE the model is fitted with linear space and time coefficients.

num_strips

The number of latitudinal/longitudinal strips to produce. This is used when plotting using autoplot.

lag

Lag can be either 0 or 1.

...

Other arguments that need to be used for datafames or currently ignored.

Examples

if (FALSE) {
library(stars)
library(dplyr)
library(units)
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
prec2 <- prec %>%
         slice(time,  1:3)
emp_spatial_cov(prec2,
                lat_or_lon_strips = "lon",
                num_strips = 4,
                lag = 1)
}