Computes spatial empirical means using a stars object as input

This function computes spatial empirical means by latitude and longitude averaged over time.

Usage

# S3 method for stars
spatial_means(x, ...)

Arguments

x

A stars object or a dataframe. Arguments differ according to the input type.

...

Other arguments that need to be used for datafames or currently ignored.

Examples

# Example 1
library(stars)
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
#> no 'var' specified, using Total_precipitation_surface_1_Hour_Accumulation
#> other available variables:
#>  lat, lon, time
#> Will return stars object with 236118 cells.
#> No projection information found in nc file. 
#>  Coordinate variable units found to be degrees, 
#>  assuming WGS84 Lat/Lon.
sp_means <- spatial_means(prec)
sp_means
#> $spatial_avg
#> # A tibble: 10,266 × 3
#> # Groups:   x [10,215]
#>        x     y   mu_emp
#>    <dbl> <dbl> [kg/m^2]
#>  1 -80.6  33.8        0
#>  2 -80.6  33.8        0
#>  3 -80.6  33.8        0
#>  4 -80.6  33.8        0
#>  5 -80.6  33.9        0
#>  6 -80.6  33.8        0
#>  7 -80.5  33.9        0
#>  8 -80.5  33.8        0
#>  9 -80.5  33.8        0
#> 10 -80.5  33.9        0
#> # … with 10,256 more rows
#> 
#> $data
#> # A tibble: 236,118 × 4
#>        x     y t                      value
#>    <dbl> <dbl> <dttm>              [kg/m^2]
#>  1 -80.6  33.8 2018-09-13 19:00:00        0
#>  2 -80.6  33.8 2018-09-13 19:00:00        0
#>  3 -80.5  33.8 2018-09-13 19:00:00        0
#>  4 -80.5  33.7 2018-09-13 19:00:00        0
#>  5 -80.5  33.7 2018-09-13 19:00:00        0
#>  6 -80.4  33.7 2018-09-13 19:00:00        0
#>  7 -80.4  33.7 2018-09-13 19:00:00        0
#>  8 -80.3  33.7 2018-09-13 19:00:00        0
#>  9 -80.3  33.7 2018-09-13 19:00:00        0
#> 10 -80.3  33.6 2018-09-13 19:00:00        0
#> # … with 236,108 more rows
#> 
#> $data_stars
#> stars object with 2 dimensions and 1 attribute
#> attribute(s):
#>                                    Min.   1st Qu.   Median     Mean  3rd Qu.
#> Total_precipitation_surface_1_...     0 0.2834783 1.588261 4.143009 6.505326
#>                                        Max.
#> Total_precipitation_surface_1_...  27.60565
#> dimension(s):
#>   from  to refsys                         values x/y
#> x    1  87 WGS 84 [87x118] -80.6113,...,-74.8822 [x]
#> y    1 118 WGS 84   [87x118] 32.4413,...,37.6193 [y]
#> curvilinear grid
#> 
#> $call
#> spatial_means.stars(x = prec)
#> 
#> attr(,"class")
#> [1] "spatialmeans"

# Example 2
tif = system.file("tif/olinda_dem_utm25s.tif", package = "stars")
x <- read_stars(tif)
sp_means <- spatial_means(x)
#> Only one observation of each location is present. The mean is the original value itself!
sp_means
#> $spatial_avg
#> # A tibble: 12,321 × 3
#> # Groups:   x [111]
#>          x        y mu_emp
#>      <dbl>    <dbl>  <dbl>
#>  1 288821. 9110816.     12
#>  2 288821. 9110906.     19
#>  3 288821. 9110996.     14
#>  4 288821. 9111086.     16
#>  5 288821. 9111176.      9
#>  6 288821. 9111266.      9
#>  7 288821. 9111356.     14
#>  8 288821. 9111446.     14
#>  9 288821. 9111536.     11
#> 10 288821. 9111626.      9
#> # … with 12,311 more rows
#> 
#> $data
#> # A tibble: 12,321 × 3
#>          x        y value
#>      <dbl>    <dbl> <dbl>
#>  1 288821. 9120716.    38
#>  2 288911. 9120716.    49
#>  3 289001. 9120716.    64
#>  4 289091. 9120716.    65
#>  5 289181. 9120716.    69
#>  6 289271. 9120716.    55
#>  7 289361. 9120716.    38
#>  8 289451. 9120716.    63
#>  9 289541. 9120716.    69
#> 10 289631. 9120716.    59
#> # … with 12,311 more rows
#> 
#> $data_stars
#> stars object with 2 dimensions and 1 attribute
#> attribute(s):
#>                        Min. 1st Qu. Median     Mean 3rd Qu. Max.
#> olinda_dem_utm25s.tif    -1       6     12 21.66521      35   88
#> dimension(s):
#>   from  to  offset    delta                       refsys point x/y
#> x    1 111  288776  89.9941 UTM Zone 25, Southern Hem... FALSE [x]
#> y    1 111 9120761 -89.9941 UTM Zone 25, Southern Hem... FALSE [y]
#> 
#> $call
#> spatial_means.stars(x = x)
#> 
#> attr(,"class")
#> [1] "spatialmeans"