Computes temporal empirical means using a stars object as input

This function computes temporal empirical means averaged per time unit.

Usage

# S3 method for stars
temporal_means(x, ...)

Arguments

x

A stars object or a dataframe. Arguments differ according to the input type.

...

Other arguments currently ignored.

Examples

library(stars)
library(dplyr)
library(units)
# Example
prec_file = system.file("nc/test_stageiv_xyt.nc", package = "stars")
prec <- read_ncdf(prec_file)
#> no 'var' specified, using Total_precipitation_surface_1_Hour_Accumulation
#> other available variables:
#>  lat, lon, time
#> Will return stars object with 236118 cells.
#> No projection information found in nc file. 
#>  Coordinate variable units found to be degrees, 
#>  assuming WGS84 Lat/Lon.
temporal_means(prec)
#> $data
#> # A tibble: 236,118 × 3
#>       value time                id                                
#>    [kg/m^2] <dttm>              <chr>                             
#>  1        0 2018-09-13 19:00:00 -80.6112976074219-33.781177520752 
#>  2        0 2018-09-13 19:00:00 -80.5721740722656-33.7664184570312
#>  3        0 2018-09-13 19:00:00 -80.533073425293-33.7516403198242 
#>  4        0 2018-09-13 19:00:00 -80.4939956665039-33.7368392944336
#>  5        0 2018-09-13 19:00:00 -80.4549407958984-33.7220191955566
#>  6        0 2018-09-13 19:00:00 -80.4159088134766-33.7071800231934
#>  7        0 2018-09-13 19:00:00 -80.3769073486328-33.6923217773438
#>  8        0 2018-09-13 19:00:00 -80.3379287719727-33.6774444580078
#>  9        0 2018-09-13 19:00:00 -80.2989730834961-33.6625442504883
#> 10        0 2018-09-13 19:00:00 -80.2600402832031-33.6476249694824
#> # … with 236,108 more rows
#> 
#> $averages
#> # A tibble: 23 × 2
#>    time                   meanz
#>    <dttm>              [kg/m^2]
#>  1 2018-09-13 19:00:00     2.40
#>  2 2018-09-13 20:00:00     2.79
#>  3 2018-09-13 21:00:00     2.30
#>  4 2018-09-13 22:00:00     2.05
#>  5 2018-09-13 23:00:00     3.21
#>  6 2018-09-14 00:00:00     4.35
#>  7 2018-09-14 01:00:00     3.57
#>  8 2018-09-14 02:00:00     4.11
#>  9 2018-09-14 03:00:00     3.84
#> 10 2018-09-14 04:00:00     4.96
#> # … with 13 more rows
#> 
#> $call
#> temporal_means.stars(x = prec)
#> 
#> attr(,"class")
#> [1] "temporalmeans"